Update admin_card.php
A bug in deleting a card has been repaired. 
It used to be on the basis of existence in table borrow to decide whether it shall be removed. However, the borrow record won't be deleted from table even after the user has returned a book. Sequentially, not only will the card tuple be deleted from borrow, but all corresponding records in borrow will be deleted since it has a foreign key referring to the borrow.